+++ src/input/xi.c
@@ -4,14 +4,14 @@
 extern Display* g_display;
 extern Window g_window;
 
-int input_init() {
+uint8_t input_init() {
    XSelectInput( g_display, g_window, ExposureMask | KeyPressMask );
    return 1;
 }
 
-int input_poll() {
+uint8_t input_poll() {
    XEvent event;
-   int mask = 0;
+   uint16_t mask = 0;
 
    mask = ExposureMask | KeyPressMask | KeyReleaseMask;
